[WARNING] Removed pointer to controls_t in joystick_parsing
The module joystick parsing was using a pointer to a control_t in
central data (named joystick_controls) to store incoming channels. A pointer was not
required and not all the fields of controls_t were used. Now the
channels are stored internally in the field channels of
joystick_parsing_t.

CHANGES IN MODULE INITIALIZATION:
//--------------------------------------------------------------------------
// Init joystick
joystick_parsing_init(	&central_data.joystick_parsing,
			&central_data.state );

Also in central_data_t, joystick_controls is no longer required.